FInished transfer function, replaced double with decimal for higher precision, converted solana and etherum wallets to doublewallet logic (double as in fungible anda non fungible tokens)

+++ CrytpWallet/Classes/Assets/Asset.cs
@@ -11,8 +11,8 @@ namespace CrytpWallet.Assets
     {
         public Guid Adress { get; init;  }
         public string Name { get; init; }
-        public double ValueInDollar { get; protected set; }
-        public double OldValueInDollar { get; protected set; }
+        public decimal ValueInDollar { get; protected set; }
+        public decimal OldValueInDollar { get; protected set; }
         public virtual void PrintAsset()
         {
             Console.WriteLine($"Adresa asseta: {Adress}\n" +
